Refactor: extract slider from volume bar
