+++ src/components/VolumeBar.tsx
@@ -1,6 +1,6 @@
 import { useUserAgent } from "@oieduardorabelo/use-user-agent";
-import { FC, useEffect, useRef, useState } from "react";
-
+import { FC } from "react";
+import { Slider } from "./Slider";
 
 export const VolumeBar: FC<{
   volume: number;
@@ -10,8 +10,6 @@ export const VolumeBar: FC<{
 
   const isIOS = ua?.os?.name === "iOS";
 
-  const [clicked, setClicked] = useState(false);
-
   const iconClass =
     volume >= 70
       ? "fa-volume-up"
@@ -19,83 +17,10 @@ export const VolumeBar: FC<{
       ? "fa-volume-down"
       : "fa-volume-off";
 
-  const { dragRef, getDrag } = useDrag((p) => setVolume(p), clicked);
-
   return isIOS ? null : (
     <div className="section" id="volume">
       <i className={`fa ${iconClass}`}></i>
-      <div
-        className="progressbar"
-        ref={dragRef}
-        onMouseDown={(e) => {
-          setClicked(true);
-          dragRef.current &&
-            setVolume(
-              getDrag({ kind: "mouse", e: e.nativeEvent }, dragRef.current)
-            );
-        }}
-        onMouseUp={() => setClicked(false)}
-      >
-        <div className="line">
-          <div
-            className={clicked ? "inside clicked" : "inside"}
-            style={{
-              width: `${volume}%`,
-            }}
-          ></div>
-        </div>
-      </div>
+      <Slider percentage={volume} setPercentage={(v) => setVolume(v)}></Slider>
     </div>
   );
-};
-
-const useDrag = (onDrag: (percent: number) => void, clicked: boolean) => {
-  const draggable = useRef<HTMLDivElement>(null);
-
-  type VolumeTrigger =
-    | {
-        kind: "mouse";
-        e: MouseEvent;
-      }
-    | {
-        kind: "touch";
-        e: TouchEvent;
-      };
-
-  const getDrag = (trigger: VolumeTrigger, bar: HTMLElement) => {
-    const horizontalOffsetPx =
-      trigger.kind == "mouse"
-        ? trigger.e.offsetX ||
-          trigger.e.clientX - bar.getBoundingClientRect().left ||
-          0
-        : trigger.e.changedTouches[0].clientX -
-            bar.getBoundingClientRect().left || 0;
-
-    return 100 * Math.min(Math.max(0, horizontalOffsetPx) / bar.offsetWidth, 1);
-  };
-
-  useEffect(() => {
-    const bar = draggable.current;
-    if (!bar) return;
-    const onmove = (e: VolumeTrigger) => {
-      if (e.kind == "touch") {
-        e.e.preventDefault();
-        onDrag(getDrag(e, bar));
-      } else if (clicked) {
-        onDrag(getDrag(e, bar));
-      }
-    };
-
-    const onmousemove = (e: MouseEvent) => onmove({ e, kind: "mouse" });
-    const ontouchmove = (e: TouchEvent) => onmove({ e, kind: "touch" });
-
-    bar.addEventListener("mousemove", onmousemove);
-    bar.addEventListener("touchmove", ontouchmove);
-    return () => {
-      bar.removeEventListener("touchmove", ontouchmove);
-      bar.removeEventListener("mousemove", onmousemove);
-    };
-  }, [draggable.current, onDrag, clicked]);
-
-  return { dragRef: draggable, getDrag: getDrag };
 };
